+++ CHANGELOG.md
@@ -6,6 +6,14 @@
 
 - Added the option to set `discord_token` in the plugins' config.
 
+### Changed
+
+- The bot no longer adds ">" in front of messages.
+
+### Fixed
+
+- Fixed a bug causing the Bot to not disconnect on plugin shutdown.
+
 ### Removed
 
 - `discord_token` can no longer be set as an environment variable.